chore: move toml handling to own module
